requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
flask==2.3.1
itsdangerous==2.1.2
dnspython==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.6.0
flask-jwt-extended==4.4.4